requirements.txt
PySocks
python-telegram-bot==13.15
requests
paramiko
